Update with cleaner theme library files
